backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
PyMySQL==1.1.1
cryptography==44.0.1
python-dotenv==1.0.1
pytest==8.3.5
